[isoltest] Moves error types to dedicated file.
