+++ test/TestCase.h
@@ -35,14 +35,6 @@ namespace solidity
 namespace test
 {
 
-#define soltestAssert(CONDITION, DESCRIPTION) \
-	do \
-	{ \
-		if (!(CONDITION)) \
-			BOOST_THROW_EXCEPTION(runtime_error(DESCRIPTION)); \
-	} \
-	while (false)
-
 /**
  * Common superclass of anything that can be run via isoltest.
  */
